requirements.txt
# Web framework
flask>=2.3
flask-mail>=0.9.1
flask-mysqldb>=1.0.1

# ML & data
scikit-learn>=1.3
numpy>=1.24
pandas>=2.0
joblib>=1.3
imbalanced-learn>=0.11
xgboost>=2.0

# Notebooks
jupyter>=1.0
ipykernel>=6.29
matplotlib>=3.7
seaborn>=0.12
matplotlib-inline>=0.1.6

# Config
python-dotenv>=1.0

# Production server
gunicorn>=21.0weasyprint>=60.0
